(global) : big refactoring add anne scolaire trait

+++ src/Bundle/Admin/Controller/SkInformationController.php
@@ -3,7 +3,7 @@
  * Created by PhpStorm.
  * User: vony
  * Date: 3/30/19
- * Time: 1:31 PM
+ * Time: 1:31 PM.
  */
 
 namespace App\Bundle\Admin\Controller;
@@ -15,17 +15,11 @@ use Symfony\Component\HttpFoundation\Request;
 
 class SkInformationController extends Controller
 {
-    /**
-     *
-     */
     public function getUserConnected()
     {
         return $this->get('security.token_storage')->getToken()->getUser();
     }
 
-    /**
-     *
-     */
     public function getEntityService()
     {
         return $this->get('sk.repository.entity');
@@ -33,21 +27,24 @@ class SkInformationController extends Controller
 
     /**
      * @return \Symfony\Component\HttpFoundation\Response
+     *
+     * @throws \Exception
      */
     public function indexAction()
     {
-        $_user_ets = $this->getUserConnected()->getEtsNom();
-
         $_user_information = $this->getEntityService()->getAllListByEts(SkInformation::class);
 
         return $this->render('@Admin/SkInformation/index.html.twig', array(
-            'information' => $_user_information
+            'information' => $_user_information,
         ));
     }
 
     /**
      * @param Request $request
+     *
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     *
+     * @throws \Exception
      */
     public function newAction(Request $request)
     {
@@ -58,32 +55,34 @@ class SkInformationController extends Controller
             return $this->redirectToRoute('sk_login');
         }
         $_new_info = new SkInformation();
-        $_user_ets = $this->getUserConnected()->getEtsNom();
         $_form = $this->createForm(SkInformationType::class, $_new_info);
         $_form->handleRequest($request);
 
         if ($_form->isSubmitted() && $_form->isValid()) {
             $_new_info->setDateAjout(new \DateTime('now'));
-            $_new_info->setEtsNom($_user_ets);
             try {
                 $this->getEntityService()->saveEntity($_new_info, 'new');
                 $this->getEntityService()->setFlash('success', 'Ajout d\'information effectuée');
             } catch (\Exception $exception) {
                 $this->getEntityService()->setFlash('error', 'Une erreur s\'est produite, veuillez réessayez');
             }
+
             return $this->redirectToRoute('info_index');
         }
 
         return $this->render('AdminBundle:SkInformation:add.html.twig', array(
-            'form' =>$_form->createView(),
-            'information'=>$_new_info
+            'form' => $_form->createView(),
+            'information' => $_new_info,
         ));
     }
 
     /**
-     * @param Request $request
+     * @param Request       $request
      * @param SkInformation $information
+     *
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     *
+     * @throws \Exception
      */
     public function updateAction(Request $request, SkInformation $information)
     {
@@ -108,17 +107,27 @@ class SkInformationController extends Controller
         }
 
         return $this->render('@Admin/SkInformation/edit.html.twig', array(
-            'form'=>$_form->createView(),
-            'information'=>$information
+            'form' => $_form->createView(),
+            'information' => $information,
         ));
     }
 
     /**
      * @param SkInformation $skInformation
+     *
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
      */
     public function deleteAction(SkInformation $skInformation)
     {
+        /*
+         * Secure to etudiant connected
+         */
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_ETUDIANT')) {
+            return $this->redirectToRoute('sk_login');
+        }
         if (true === $this->getEntityService()->deleteEntity($skInformation, '')) {
             return $this->redirectToRoute('info_index');
         }
